+++ BetterChests/Framework/Models/Events/InventoryMenuChangedEventArgs.cs
@@ -1,0 +1,8 @@
+namespace StardewMods.BetterChests.Framework.Models.Events;
+
+/// <summary>Represents the event arguments for changes in the inventory menu.</summary>
+internal sealed class InventoryMenuChangedEventArgs : EventArgs
+{
+    /// <summary>Initializes a new instance of the <see cref="InventoryMenuChangedEventArgs" /> class.</summary>
+    public InventoryMenuChangedEventArgs() { }
+}